CAUSEWAY-3663: refactors PendingParamsSnapshot to be closer to ParameterNegotiationModel
